requirements.txt
Django==4.2.7
Pillow==10.1.0
django-ckeditor==6.7.0

